+++ server/src/routes/transcribeRoutes.ts
@@ -1,0 +1,8 @@
+import express from "express";
+import { handleTranscribe } from "../controllers/transcribeController";
+
+const router = express.Router();
+
+router.post("/transcribe", handleTranscribe);
+
+export default router;
